extract shared CDS helpers into cds_helper.py
